+++ client/js/init.js
@@ -1,1 +1,0 @@
-globalThis.darkMode = $.dark